assets/src/components/propertiesPanel/minischedule.tsx
import React from "react"
import {
  useMinischeduleBlock,
  useMinischeduleRun,
} from "../../hooks/useMinischedule"
import { Block, Run } from "../../minischedule"
import { TripId } from "../../schedule"
import Loading from "../loading"

export interface Props {
  activeTripId: TripId
}
export const MinischeduleRun = ({ activeTripId }: Props) => {
  const run: Run | null | undefined = useMinischeduleRun(activeTripId)
  if (run === undefined) {
    return <Loading />
  } else if (run === null) {
    return <>No run found</>
  } else {
    return <div>{JSON.stringify(run)}</div>
  }
}

export const MinischeduleBlock = ({ activeTripId }: Props) => {
  const block: Block | null | undefined = useMinischeduleBlock(activeTripId)
  if (block === undefined) {
    return <Loading />
  } else if (block === null) {
    return <>No block found</>
  } else {
    return <div>{JSON.stringify(block)}</div>
  }
}
